backend/pyproject.toml
[project]
name = "todo-web-backend"
version = "0.1.0"
description = "FastAPI backend for multi-user Todo application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "psycopg2-binary",
    "python-jose[cryptography]",
    "python-dotenv",
    "pytest",
]

[project.scripts]
todo-api = "todo.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["../tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/todo"]
